pyproject.toml
[tool.poetry]
name = "overpass"
version = "0.1.0"
description = ""
authors = ["GOATS2K <jp@glattetre.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-login = "^0.5.0"
flask-discord = "^0.1.63"
pyopenssl = "^19.1.0"
python-dotenv = "^0.15.0"
flask-wtf = "^0.14.3"
gunicorn = "^20.0.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
